Count number of candidate evaluated (#44)
In order to compare different algorithms, fixing the number of
candidates evaluated provided for a fairer comparison than using a
timeout because it allows comparisons to be done on different
machines (provided they use different GPUs), and relax the constraint
of having the machine unused during benchmarking. Since low execution
time is the final objective, we still track it, notably to ensure that
algorithms don't get too slow at producing candidates. However, even
in that case, it is arguably easier to improve an efficient algorithm
by making it faster rather than directly finding a fast and efficient
algorithm.