Count the evaluated candidates in the monitor (#47)
Previously we were counting the number of evaluated candidates in the
context through a callback parameter. Since 03526bcf, we are also
counting the number of evaluated candidates directly in the monitor in
order to stop evaluation after a possible maximum number of
candidates.

We should have a single place where we count that number of evaluated
candidates, and that place should be in the monitor because it doesn't
make sense to require each context to manually count the number of
evaluated candidates. In addition, this makes it easier to instantiate
multiple evaluators in the future if we have several identical
accelerators and let the monitor take care of the synchronization and
counting problems.