Introduce bare-bones C and Python API
This adds a new telamon-capi crate exporting a C API through a dynamic
library, and a Python package which uses said C API crate to expose a
Python API for Telamon. For the moment the API is very simple and only
allows optimising a MatMul kernel on either an x86 or cuda device, but
should evolve over time to incorporate additional features such as
driving the exploration from Python.

Fixes #49