Add ability to configure matmul tilings from the Python API
This allows manually overriding the tiling configuration for the MatMul
kernel through parameters in MatMulP and exposes that configuration
through the C and Python API.

In addition, a new Python example script is included which runs an
optimal candidate search from Python with a couple options (e.g. kernel
and device selection) which can be used as a starting point for further
experiments.