Avoid use-after-free errors in the Python API
When calling in the C API from Python we are allocating Tiling objects
from Python. The lifetime of the corresponding memory is linked to the
lifetime of the corresponding Python object. We were using temporaries
for the Python object, which goes out of scope and deallocate the memory
as soon as they are passed to the C API function. Instead, we now use
explicit variables which live until after the C API call has returned.