Ensure telamon-gen output is deterministic
We want telamon-gen output to be deterministic on a given .exh file
because we want to be able to assume field and enum order when
serializing to compact binary formats. Currently telamon-gen is
deterministic (thanks to using the non-random HashMaps from
telamon-utils) up to the printer, where we use a topological sort from
the `pathfinding` crate to ensure counters are serialized after their
dependencies. Unfortunately, the topological sort in the `pathfinding`
crate is non-deterministic due to its use of regular Rust HashMaps.

This patch replaces the topological sort from the `pathfinding` crate
with our own topological sort implementation,
`stable_topological_sort`, which we guarantee is not only
non-deterministic but is also stable: the order between elements which
don't have (transitive) dependencies to each other is preserved.

In addition, it replaces the use of some HashMaps with IndexMaps from
the `indexmap` crate, which are hash tables ensuring that iteration
order is identical to insertion order until any value is removed. This
allows to help propagate the stability guarantee to the order in which
choices are defined in the `.exh` file, giving some control on the
generated code to the user.