Enable crate expansion
This removes the integration with `build.rs` for generating the C API
headers. The C API headers must now be generated using a manual call to
`cbindgen`.

weurohis allows us to properly macro-expand crates (including the current
crate, in the future) before going through `cbindgen`: expansion
requires a nightly compiler (due to use of the `--pretty` option), and
hence is not compatible with calling `cbindgen` from a stable
`build.rs`. It also makes more sense, conceptually: the Rust build
process do not need those C headers; they are only needed by external
build processes depending on the C interface. Those external build
processes are responsible for ensuring (or not) that the C header they
use is properly up-to-date.