Pre-allocate IDs for lowering
Currently telamon is dynamically creating new instruction and dimension
IDs when performing a lowering, which means that:

 - Depending on previous choices, the same ID can be used for different
   choices semantically.
 - The total number of possible choices is not known until all the
   lowerings have been performed (or not).

Both of those properties make it awkward to build statistical models on
top of the information present in the domain. This PR addresses those
issues by pre-allocating IDs for the instruction and dimensions that
will need to be created when a dimension map is lowered.

That information is stored on a global `DimMapScope` associated with a
lowerable operand (currently only `Inst`), and is annotated on the
creation of the `SearchSpace`. We parameterize the `DimMapScope` and
dependent types (`Operator`, `Operand`, `Instruction`, `Function`, ...)
in two versions, one with and one without the associated lowering
information. The `Builder` uses the non-annotated version, then converts
the representation to the annotated version upon creation of the
`SearchSpace`.