Fix two bugs in telamon-gen.
1. Correctly run the filters when allocating new objects, even if the
filtered choice was already existing. This involves creating and
populating a structure to store such filters and generating code for
the filter calls. The printing part is awkward because the old template
system we used to generated code is only partially removed, so one part of
the printing uses rust Ast while the other use templates compiled to
strings.

2. Move some set constraints outside the forall loops nest when reverse
constraints depends on them.

This PR also makes a few small changes on the side:
- Fix a small mistake in `fail.rs`. The changes made it fail to compile.
- Fix `ir_gen`. A variable was wrongly named, but this macro case was never
  used before.