Groundwork to support DMA and dynamic layouts.
This change exposes memory layouts to the search space. A memory layout
has a set of dimensions. A `rank` variable specifies the order or the
layout dimensions. It is currently fixed but will ater be dynamic.

This change then express vectorization constraints in term of `rank`. It
also adds a test in telamon-gen and fixes minor errors in code paths
that were unused before.

Change-Id: Ib0699af43437e0c73401ee07436b648e89152a85