Add DMA instructions.
This changes adds support for DMA instructions in Telamon. For that it
adds two instructions (DmaStart and DmaWait) and a new type of
variables (SyncFlag). DmaStart produces a SyncFlag variables consumed by
DmaWait. Constraints on the SyncFlag variable ensure the start and wait
instruction have the same vectorization pattern.

Suggested review order:
- src/ir/operator.rs
- src/ir/instruction.rs
- src/ir/types.rs
- src/search_space/variable.exh
- src/codegen/function.rs
- src/codegen/printer.rs
- src/helper/builder.rs

Change-Id: I56fdf7d542261e4f33d65dda67713f41c5914fdc