Make ChoiceOrdering more idiomatic
This patch allows displaying and parsing ChoiceOrdering values, which
paves the way for using them as command line parameters.  The patch also
removes the ChoiceGroup definition in choice.rs to keep only the ones in
config.rs, greatly diminishing the bookkeeping required.  The
distinction was meant to allow more generic selection of groups at the
config level which would require extra computation to be lowered (e.g.
mapping of loop variables to internal dimension IDs), but we ended up
not doing this for the foreseeable future.

I also finally understood the proper way of using IntoIterator to write
a generic `list` version, so the patch folds `choice::list_with_conf`
into the generi `choice::list`.

telamon:
 * src/explorer/config.rs:
	Add FromStr, Display and IntoIterator implementations.
 * src/explorer/choice.rs:
	Use config::ChoiceGroup and the new IntoIterator directly.
 * src/explorer/local_selection.rs:
	Use `choice::list` instead of `choice::list_with_conf`.
 * src/explorer/bandit_arm.rs:
	Use `choice::list` instead of `choice::list_with_conf`.