Add object-safe trait for building candidates
This patch introduces a `KernelBuilder` structure which wraps a kernel's
parameters.  The `KernelBuilder` provides a `scoped_kernel` helper
function which takes care of building the kernel itself and its
signature, and passes them on to a scoped closure[1].  Moreover, it
introduces a `ScopedCandidates` trait (and implements it for the
`KernelBuilder`) that allows building candidates from a type which
doesn't depend on the actual kernel used:

```rust
let builder: Box<dyn ScopedCandidates<'_, _, _, _>> = match KernelType {
	Axpy(params) => Box::new(linalg::Axpy<'_, f32>::builder(params)),
	MatMul(params) => Box::new(
		linalg::MatMul<'_, f32>::builder(params),
	)
}
```

This makes it easier to write kernel-independent tools around the core
Telamon library.

1: The closure is required to ensure lifetime invariants relative to the
   mutable reference on the execution context.

telamon-kernels/
 * src/kernel.rs:
	Introduce the `KernelBuilder` and `ScopedCandidates`, and rewrite
	the `Kernel` helper functions (`benchmark`, etc.) to use them.
	Also introduce and use a more explicit `MemInit` enum for specifying
	memory initialization procedure instead of relying on a
	`random_fill` boolean.
 * src/lib.rs:
 * benches/cuda_bound.rs:
 * benches/cuda_search.rs:
 * benches/cuda_variance.rs:

telamon-capi/
 * telamon-capi/src/lib.rs