Fix telamon-gen cc_tests
This required removing support for rustfmt throught the old `rustfmt`
crate which doesn't support rust 2018 (apparently).  Instead, we use the
`rustfmt` executable in the PATH when available.

Note that this removes support for formatting stdout streams, but that
behavior can be recovered by manually piping through `rustfmt`.

Also note that the `kernels` crate is still 2015.