Check for timeout even in deadend hell
When the tree search threads get stuck in a series of deadend without
finding implementations, the only way for them to get out of there is
for the `stop` field of the tree to be set to `true`.  This happens when
the monitor calls `stop_evaluation` after a timeout on its future.

This should work, but it seems like in some situations it doesn't.  But
in order to avoid creating gigabytes of log files, this patch adds an
additional safeguard by also checking the timeout from the tree search
directly.