Re-enable clippy checks
Clippy has gone a long way since #130, including working with stable
Rust and being less pedantic by default.  This patch makes Telamon
conform to the clippy lints (with some locally disabled checks that did
not seem appropriate), and re-enables clippy checking on Travis.

Clippy lints are skipped on telamon-gen due to the large amount of
warnings; they can be re-introduced slowly (e.g. file-by-file or
lint-by-lint) -- or not, depending on the future of telamon-gen.