Make new MCTS implementation the default
This patch makes the new MCTS implementation (in explorer::mcts) the
default over the previous explorer::bandit_arm.  explorer::bandit_arm is
kept for the moment for backwards compatibility, although the behavior
of the two implementations should be fairly similar.

The tests are now run using the new MCTS implementation, which allowed
to spot a couple of bugs -- namely, that the new implementation was not
properly detecting tree exhaustion and was not stopping in that case.
The patch also fixes that issue by marking nodes that were already
evaluated as dead, ensuring they won't be evaluated again in the future.
A little bit of thought may be needed to understand what that entails on
the theoretical "guarantees" of the MCTS algorithm.