Remove references from ir::Function
This patch removes the lifetime in `ir::Function` due to holding
references to the signature, parameters, and device.  The references are
replaced with atomically reference-counted pointers instead.  Since
those internal structures are rarely copied (they are already behind
other `Arc`s which get themselves cloned, except when during lowering),
there should be virtually no performance difference, but it makes
various APIs nicer -- and limits the lifetime hell we would occasionally
find ourselves in.

The PR touches a large amount of files, but most of the changes in there
are fairly straightforward -- it's just removing lifetimes and
(sometimes) adding `Arc`s.