Fix constraint on inner block size
The inner block dimensions are limited to a maximum of  65535 in CUDA
architectures, which we represent using a constraint.  However that
constraint had a rounding bug due to using integer division, resulting
in possibly larger block dimensions in certain cases.  This patch fixes
that bug by properly rounding upwards instead.