Split the explorer in a separate crate
This patch extracts the `telamon::explorer` module into a separate
`telamon_explorer` crate, responsible for everything related to the
search algorithms.

This should help with compilation times since modifying the explorer
will no longer require recompiling the rest of the `telamon` crate (and
in particular not the generated `choices.rs`).