Use relative error in kernel error checking
Currently we are checking kernel output using a single threshold of
`1e-5` on the absolute error.  However, for large kernels, this is not
appropriate: errors accumulate on large sums, and can exceed the
threshold even for correct computations.

Instead, this patch makes it so that we check correctness using a
combination of an absolute and relative threshold.  This alleviates the
problem since large sums will have higher absolute values¹ and hence
will allow for more leeway thanks to the relative threshold term.

The implementation is similar to the NumPy `np.allclose` function, that
is, we add the absolute and relative tolerance and compare with the
absolute difference.  The absolute and relative tolerance are 1e-8 and
1e-5 respectively, following again the default values from
`np.allclose`.

At the same time, the patch also changes the error displayed in case of
correctness errors:  instead of dumping sometimes large tensors in
a textual form, it now prints summary statistics of the absolute and
relative errors, as well as the proportion of elements failing the
`allclose` check.

Note that those two issues are related to now checking correctness of
full kernels during evaluation: where previously we were only checking
correctness on small toy examples in tests, we are now checking
correctness on large tensors -- which makes the error messages way too
large (we search axpy with sizes of 2^26 -- one single error message was
gigabytes long!), and also introduces the additional precision issues.

Additionally, one should note that, due to the transformations we
currently support, there should not be any error at all (in fact, if one
looks closely, the errors for all correct Telamon kernels are always
the same).  In particular, we do not reorder the arithmetic operations.
The error that we observe should thus be attributed to (a) optimizations in
the reference implementations we compare against compared to the naive
version and possibly (b) the use of fused multiply-add instructions on
some platforms.

 ¹: Except where loss of significance occurs, typically.  But since we
    only generate positive random floats this shouldn't be an issue for
    us in practice; if it is, we can revisit the approach and e.g. scale the
    error depending on the kernel sizes.  The advantage of using a
    relative error is that it doesn't require any inner knowledge of the
    kernel semantics and provide a generally acceptable check.

kernels/:
 - src/lib.rs:  Add new error-checking functions
 - src/linalg.rs:  Use the new error-checking functions
 - src/kernel.rs:  Add missing space.