Check results when exploring
Currently we are only checking correctness in specific toy examples in
the tests; however, this is limited -- because we only run those tests
on CPU, they can miss backend-specific issues, but are also fully random
(we don't have a performance model to guide the search on CPU) which can
hide some issues which would happen in practice.

This patch makes it so that correctness tests are made during the
exploration as well, either for all new best implementations found (the
default) or for all implementations (for stress-testing).  When an
implementation fails the correctness check, we handle it the same as if
it had an infinite execution time -- which is technically incorrect but
should be good enough in practice for now.

The correctness checking is implemented by passing a checking function
as argument to the explorer APIs, which works even though it is kind of
ugly (those APIs are stating to grow an long list of arguments).  A
cleaner API would require some additional thought and refactoring, so I
leave it for future work.

Finally, there was an issue with the code in the cuda backend skipping
candidates for which the bound is higher than the current best:  since
the correctness check happens at a higher level, the backend thought it
had seen very fast implementations and using them to skip a bunch of
other implementations -- but the very fast implementations were actually
incorrect, and no longer used by the explorer.  Instead, I lifted the
skipping code to the explorer, resulting in an API change to the
`AsyncCallback` which now takes an `ExecutableKernel` as argument,
giving it control late in the pipeline to execute the kernel or not.  In
the same go I moved the stabilization from the CUDA crate to the
explorer as well, even though I realize when writing these lines that it
was not necessary.  Oh well, at least it is done now.

Most of the changes correspond to simple updates to the new APIs; the
crux of the work happen in `device/context.rs` and `explorer/mod.rs`.

Important changes:
	backend/cuda/:
		src/context.rs:  API update; remove stabilization
	backend/mppa/:
		src/context.rs:  API update; timing measurements fix
	backend/x86/:
		src/context.rs:  API update
	src/
		device/context.rs:  New `KernelEvaluator` and `Stabilizer` APIs
		explorer/candidate.rs: Allow creating with actions and dumping error data
		explorer/config.rs:  Check all or only best implementations.
		explorer/mcts.rs:  Rebuild the list of actions for the evaluator.
		explorer/mod.rs:  Check correctness and stabilize evaluations.