Unify `futures` and `crossbeam` versions
This patch makes us use the same version of `futures` and `crossbeam`
between the different crates, and avoids using the yanked 0.2 version of
futures.  Note that `tokio` is incompatible with `futures` 0.2.

It also removes the dependency on our custom patched `tokio-timer`.
Instead, we use a very simple implementation of a timeout using a
background thread which sleeps until the timeout it reached.  This is
very similar to what `tokio-timer` does under the hood -- except that
`tokio-timer` is orders of magnitude more complex because it is designed
to be able to dispatch a lot of timers at once and shares the background
thread between them.  It also uses some more precise mechanisms in
addition to `sleep` -- but we don't care about all those features: we
only ever have a single timeout in the system and a simple sleeping
thread does the job.