Derive `Ord` and `PartialOrd` for actions
This patch derives the `Ord` and `PartialOrd` trait for actions,
allowing them to be ordered and sorted (typically for display).  Since
we already use `min`, `max`, `lt`, etc. functions on the bit enums for
our own custom traits, there were confusion with the native `Ord` and
`PartialOrd` methods, so the custom ones are renamed.