Add max instruction
Add a new instruction `max`, calculating the maximum of two f32, f64, i32 or i64
values.

For the CUDA backend, the instruction is mapped directly to the max
instruction. On x86 and MPPA, the instruction is mapped to an invocation of a
new macro telamon_op_max(), returning the maximum of its two arguments.