Format floating point literals correctly in x86 and MPPA backends
Currently, both the x86 and MPPA backend print floating point literals using
f64::to_string(). This causes two problems.

First, floating point values without decimal places are printed without a
trailing ".0", which lets the C compiler interpret them as integer
literals. This results in a warning for valid floating point values which are
outside of the range of an integer.

Second, this prevents the C compiler from emitting a warning for values that
should be represented as C floats, but which are out of range.

This patch causes floating point constants to be printed as C99 hexadecimal
floating point literals with proper suffixes for floats and doubles, avoiding
the above-mentioned problems. It assumes that the targeted C implementation uses
32 bit values for floats and 64 bit values for doubles, that all f32 values can
be represented by a float and that all f64 values can be represented by a
double.

For floating point types with an unusual number of bits (i.e., neither 32 nor 64
bits), the formatting function panics.