Variable name for closed dims refer to the last iteration
The name map incorrectly assumed that it should use the *first*
iteration index instead of the *last* one when some unrolling dimensions
are not used by the underlying instruction (eg because it has been
replicated).  This makes us generate invalid code for reductions,
because it causes the reduction operand to get out of sync, generating
code similar to the following:

```
x0 = 0
x1 = 0
x0 = a0 * b0 + x1
x0 = a1 * b1 + x1
```

instead of

```
x0 = 0
x1 = 0
x1 = a0 * b0 + x1
x1 = a1 * b1 + x1
```

This is a partial fix for the same issue as #226.  We are still creating
too many variables when unrolled loops are unrolled, but at least the
generated code should be correct.