Remove x86 random correctness tests
This removes the x86 correctness tests.  We don't have a performance
model for x86, so those tests are (very) slow since we can easily end up
in incredibly bad configurations, and they don't test very meaningful
things (in the end; we only care about correctness of the best
performing code -- not in weird corner cases of the search space).  They
are also less useful than they were, since we now do correction checks
during search -- which is what matters in the end.

Note that this doesn't remove ability to perform the correctness check;
they are just not done during automated testing.  Also note that this
only concerns the random tests; the fixed tests (from `kernel_dump`) are
not removed by this patch.