Add Cargo.lock to repository
As https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
says:

> If you’re building a non-end product, such as a rust library that
> other rust packages will depend on, put Cargo.lock in your .gitignore.
> If you’re building an end product, which are executable like
> command-line tool or an application, or a system library with crate-type
> of staticlib or cdylib, check Cargo.lock into git.

Technically `telamon` (the crate) is a library, but its only users are
the `kernels` and `telamon-capi` crates, which are respectively used
through command-line tools and a system .so cdylib.  Hence there are no
good reasons not to keep the `Cargo.lock` in the repository, at least
until the day (if it comes) there are other Rust packages (outside of
this repo) which use `telamon` (the crate).

This allows to have shared, known-working dependency versions across
developers and machines.