Add test checking reproducibility of code generated by the MPPA backend
This new test reads a list of pairs each composed by a list of actions defining
an implementations of the matrix multiplication kernel from `kernels` and the
expected code to be generated by the MPPA backend for the implementation from a
file in JSON format, invokes the MPPA backend to generate the code for each
implementation and checks if the generated code matches the expected code.

A dump file can be generated by using the dump_generated_code
program. This program is currently generated as an example instead of
a binary to avoid cyclic dependencies with `kernels`. By default, the
actions and code for 20 implementations below a cut of 2e8 are dumped
to `dumps/implementation_code.json`.

Example usage:

  $ cd backend/mppa
  $ cargo run --example dump_generated_code
  $ cargo test printer_generate_code