Add a new kernel Fused2MM
Add a new kernel Fused2MM, calculating `E = alpha*A.B.C + beta*D` and
applying an activation function to each element of `E`. A, B, C and D
are 2D matrices and alpha and beta are scalars.

The activation function can either be ReLU or the Sigmoid
function. Application of the activation function can be skipped by
specifying `None` as the activation function.