Rename `DimMap::filter` into `DimMap::drain_filter`
`DimMap::filter` is a confusing name: not only does it return an
iterator over the filtered elements, but it also removes them from the
`DimMap`.  This can make for very surprising behavior, since `.filter()`
(e.g. on an iterator) usually does not have side effects.

The behavior and intent is (almost) identical to the nightly
`drain_filter` on `Vec`, so let's use that name instead to be less
confusing and more explicit.