Allow creating explorer config from any file
Previously explorer configuration could only be created from
`Settings.toml` files, or a json string.  This patch makes it so that
explorer configuration can be created by any file or json string as
needed by the caller.

In addition, we push errors to the creator of the `Config` instead of
unwrapping, and we output a warning message when automatically loading
from a `Settings.toml` since this is essentially context-dependent
defaults which can be confusing.