Allow creating explorer config from any file
Previously an explorer configuration could only be created from
`Settings.toml` files, or a json string.

This patch adds two functions from_settings_toml and from_path.  The
function from_settings_toml explicitly tries to load a configuration
from Settings.toml. If no such file is found, default parameters are
applied and a warning is generated to avoid confusion. from_path
attempts to load the configuration from an arbitrary file and reports an
error if the contents could not be loaded / parsed.