Add a telamon-cli package with CLI utilities
The telamon-cli package currently provides two executables:

 - A `cuda_search` executable which runs the search for a given kernel
   and configuration

 - A `tlcli` swiss knife executable which allows doing a bunch of handy
   and common tasks such as extracting lists of actions from event log,
   displaying eventlog statistics, replaying a list of action, etc.
   Those are scripts I had lying around that I find useful, and as such
   worth it to share.

(Yes, those should be binaries in the main `telamon` package.  They are
not because (a) it would require recompiling the main `telamon` package
each time, which is way too slow and (b) they depend on the kernels.
Yes, this should be fixed; but it requires a non negligible amount of
refactoring -- so in the meantime let's use a separate crate)