Add partial support for duplication of virtual tensors
When composing an expression using virtual tensors, it might be
necessary to reference an intermediate result more than once. For
example, when normalizing a vector, its values are referenced first by
a reduction and a second time when each element of the vector is
divided by the result of the reduction.

When simply reusing a virtual tensor in multiple subexpressions the
exact same dimensions of the virtual tensor could be used by multiple
instructions with specific ordering constraints ensuring the correct
order of the calculation, which in turn could lead to cyclic
dependencies for the ordering of the virtual tensor dimensions.

For example, in the above-mentioned normalization, the dimensions of
the virtual tensor representing the vector are used both by the
reduction and the division. For correct results, the division
instruction must be executed only once the reduction is completed,
which requires that the reduction dimension are placed before the
division. However, the division also iterates over the dimensions of
the virtual tensor. The dimensions of the virtual tensor would thus be
required to be placed before themselves and Telamon would fail with
unsatisfiable constraints.

This patch introduces a new function `duplicate()` for virtual
tensors, creating a new virtual tensor with identical values, but with
a new set of dimensions. This function is currently implemented only
partially:

  - For virtual tensors originating from a tensor stored in global
    memory, `duplicate()` simply reloads the tensor a second time from
    global memory.

  - For virtual tensors which originate from arbitrary instructions,
    duplication would require a more complex procedure, potentially
    duplicating multiple instructions or storing intermediate results
    in memory. Neither duplication of instructions, nor temporary
    buffers are currently supported by the Telamon
    infrastructure. Therefore, `duplicate()` panics in these cases.

A complete implementation covering all cases is left for a redesign of
the code for kernel composition.